A-Packages: Organize into Packages. Moved everything under package Duke
