+++ javadoc/command/Events.html
@@ -123,10 +123,10 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li><a href="Task.html" title="class in command">command.Task</a></li>
+<li><a href="Task.html" title="class in command">duke.command.Task</a></li>
 <li>
 <ul class="inheritance">
-<li>command.Events</li>
+<li>duke.command.Events</li>
 </ul>
 </li>
 </ul>
@@ -166,7 +166,7 @@ extends <a href="Task.html" title="class in command">Task</a></pre>
 </tr>
 </table>
 <ul class="blockList">
-<li class="blockList"><a id="fields.inherited.from.class.command.Task">
+<li class="blockList"><a id="fields.inherited.from.class.duke.command.Task">
 <!--   -->
 </a>
 <h3>Fields inherited from class&nbsp;command.<a href="Task.html" title="class in command">Task</a></h3>
@@ -229,7 +229,7 @@ extends <a href="Task.html" title="class in command">Task</a></pre>
 </tr>
 </table>
 <ul class="blockList">
-<li class="blockList"><a id="methods.inherited.from.class.command.Task">
+<li class="blockList"><a id="methods.inherited.from.class.duke.command.Task">
 <!--   -->
 </a>
 <h3>Methods inherited from class&nbsp;command.<a href="Task.html" title="class in command">Task</a></h3>
