+++ javadoc/command/Task.html
@@ -123,7 +123,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li>command.Task</li>
+<li>duke.command.Task</li>
 </ul>
 </li>
 </ul>
