+++ javadoc/command/ToDos.html
@@ -123,10 +123,10 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li><a href="Task.html" title="class in command">command.Task</a></li>
+<li><a href="Task.html" title="class in command">duke.command.Task</a></li>
 <li>
 <ul class="inheritance">
-<li>command.ToDos</li>
+<li>duke.command.ToDos</li>
 </ul>
 </li>
 </ul>
@@ -153,7 +153,7 @@ extends <a href="Task.html" title="class in command">Task</a></pre>
 </a>
 <h3>Field Summary</h3>
 <ul class="blockList">
-<li class="blockList"><a id="fields.inherited.from.class.command.Task">
+<li class="blockList"><a id="fields.inherited.from.class.duke.command.Task">
 <!--   -->
 </a>
 <h3>Fields inherited from class&nbsp;command.<a href="Task.html" title="class in command">Task</a></h3>
@@ -214,7 +214,7 @@ extends <a href="Task.html" title="class in command">Task</a></pre>
 </tr>
 </table>
 <ul class="blockList">
-<li class="blockList"><a id="methods.inherited.from.class.command.Task">
+<li class="blockList"><a id="methods.inherited.from.class.duke.command.Task">
 <!--   -->
 </a>
 <h3>Methods inherited from class&nbsp;command.<a href="Task.html" title="class in command">Task</a></h3>
