+++ javadoc/tool/Command.html
@@ -123,7 +123,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li>tool.Command</li>
+<li>duke.tool.Command</li>
 </ul>
 </li>
 </ul>
@@ -210,7 +210,7 @@ extends java.lang.Object</pre>
 </tr>
 <tr id="i0" class="altColor">
 <td class="colFirst"><code>boolean</code></td>
-<th class="colSecond" scope="row"><code><span class="memberNameLink"><a href="#execute(tool.TaskList,tool.UI)">execute</a></span>&#8203;(<a href="TaskList.html" title="class in tool">TaskList</a>&nbsp;taskList,
+<th class="colSecond" scope="row"><code><span class="memberNameLink"><a href="#execute(duke.tool.TaskList,tool.UI)">execute</a></span>&#8203;(<a href="TaskList.html" title="class in tool">TaskList</a>&nbsp;taskList,
        <a href="UI.html" title="class in tool">UI</a>&nbsp;ui)</code></th>
 <td class="colLast">
 <div class="block">Executes commands such as: bye, list, done, find, delete, todo, deadline, event</div>
@@ -297,7 +297,7 @@ extends java.lang.Object</pre>
 <!--   -->
 </a>
 <h3>Method Detail</h3>
-<a id="execute(tool.TaskList,tool.UI)">
+<a id="execute(duke.tool.TaskList,tool.UI)">
 <!--   -->
 </a>
 <ul class="blockListLast">
