+++ javadoc/tool/Parser.html
@@ -123,7 +123,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li>tool.Parser</li>
+<li>duke.tool.Parser</li>
 </ul>
 </li>
 </ul>
@@ -182,7 +182,7 @@ extends java.lang.Object</pre>
 <th class="colLast" scope="col">Description</th>
 </tr>
 <tr class="altColor">
-<th class="colConstructorName" scope="row"><code><span class="memberNameLink"><a href="#%3Cinit%3E(tool.UI,tool.TaskList)">Parser</a></span>&#8203;(<a href="UI.html" title="class in tool">UI</a>&nbsp;ui,
+<th class="colConstructorName" scope="row"><code><span class="memberNameLink"><a href="#%3Cinit%3E(duke.tool.UI,tool.TaskList)">Parser</a></span>&#8203;(<a href="UI.html" title="class in tool">UI</a>&nbsp;ui,
       <a href="TaskList.html" title="class in tool">TaskList</a>&nbsp;taskList)</code></th>
 <td class="colLast">&nbsp;</td>
 </tr>
@@ -263,7 +263,7 @@ extends java.lang.Object</pre>
 <!--   -->
 </a>
 <h3>Constructor Detail</h3>
-<a id="&lt;init&gt;(tool.UI,tool.TaskList)">
+<a id="&lt;init&gt;(duke.tool.UI,tool.TaskList)">
 <!--   -->
 </a>
 <ul class="blockListLast">
