+++ javadoc/tool/Storage.html
@@ -123,7 +123,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li>tool.Storage</li>
+<li>duke.tool.Storage</li>
 </ul>
 </li>
 </ul>
