+++ javadoc/tool/TaskList.html
@@ -123,7 +123,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li>tool.TaskList</li>
+<li>duke.tool.TaskList</li>
 </ul>
 </li>
 </ul>
