+++ javadoc/tool/UI.html
@@ -123,7 +123,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li>tool.UI</li>
+<li>duke.tool.UI</li>
 </ul>
 </li>
 </ul>
