+++ src/main/java/duke/command/Events.java
@@ -1,10 +1,13 @@
-package command;
+package duke.command;
 
-/** Tasks that need to be done before a specific date/time e.g., submit report by 11/10/2019 5pm */
-public class Deadlines extends Task {
+/**
+ * Tasks that start at a specific time and ends at a specific time e.g., team project meeting on
+ * 2/10/2019 2-4pm
+ */
+public class Events extends Task {
   protected String datetime = "";
 
-  public Deadlines(String description, String datetime) {
+  public Events(String description, String datetime) {
     super(description);
     this.datetime = datetime;
   }
@@ -12,9 +15,9 @@ public class Deadlines extends Task {
   /**
    * @param isDone Whether the task is completed by the user
    * @param description The activity description
-   * @param datetime The date and time described in /by command
+   * @param datetime The date and time described in /at command
    */
-  public Deadlines(boolean isDone, String description, String datetime) {
+  public Events(boolean isDone, String description, String datetime) {
     super(description);
     this.isDone = isDone;
     this.datetime = datetime;
@@ -22,13 +25,13 @@ public class Deadlines extends Task {
 
   @Override
   public String toString() {
-    return "[D]" + "[" + getStatusIcon() + "] " + this.description + " (by: " + this.datetime + ")";
+    return "[E]" + "[" + getStatusIcon() + "] " + this.description + " (at: " + this.datetime + ")";
   }
 
   /** @return String to be stored in task text file */
   @Override
   public String fileString() {
-    return "D|" + getStatusIcon() + "|" + this.description + "|" + this.datetime;
+    return "E|" + getStatusIcon() + "|" + this.description + "|" + this.datetime;
   }
 
   @Override
